Finish first placement of divs
Just made a few divs to more or less know how those divs will be placed
when the user chooses how many of those the container div will hold. I
created a few variables referring to the size of the container grid, the
number of divs, the gap between divs... just to make my life easier in
the future.